USRP Now Works Using Python (Kinda)
The USRP now works using Python, some documentation is attached explaining setup and what to do. Threading needs to be implemented to allow transmitting and receiving to occur in tandem with each other. Hopefully we can get access to the X410 again next semester or the X310 to finish the threading issue.